+++ Tests/Test-CSmbShare.Tests.ps1
@@ -25,20 +25,20 @@ BeforeAll {
     Set-StrictMode -Version 'Latest'
 
     & (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-CarbonTest.ps1' -Resolve)
-    $script:shareName = 'CarbonTestFileShare'
+    $script:shareName = 'CarbonTestSmbShare'
     $script:sharePath = $null
     $script:shareDescription = 'Share for testing Carbon''s Get-FileShare function.'
 
     $script:sharePath = New-CTempDirectory -Prefix $PSCommandPath
-    Install-CFileShare -Path $script:sharePath -Name $script:shareName -Description $script:shareDescription
+    Install-CSmbShare -Path $script:sharePath -Name $script:shareName -Description $script:shareDescription
 }
 
 AfterAll {
-    Uninstall-CFileShare -Name $script:shareName
+    Uninstall-CSmbShare -Name $script:shareName
     Uninstall-CDirectory -Path $script:sharePath
 }
 
-Describe 'Test-FileShare' {
+Describe 'Test-CSmbShare' {
     BeforeEach {
         $Global:Error.Clear()
     }
@@ -46,13 +46,13 @@ Describe 'Test-FileShare' {
     It 'should test share' {
         $shares = Get-CFileShare
         $shares | Should -Not -BeNullOrEmpty
-        $sharesNotFound = $shares | Where-Object { -not (Test-CFileShare -Name $_.Name) }
+        $sharesNotFound = $shares | Where-Object { -not (Test-CSmbShare -Name $_.Name) }
         $sharesNotFound | Should -BeNullOrEmpty
         $Global:Error | Should -BeNullOrEmpty
     }
 
     It 'should detect shares that do not exist' {
-        (Test-CFileShare -Name 'fdjfkdsfjdsf') | Should -BeFalse
+        (Test-CSmbShare -Name 'fdjfkdsfjdsf') | Should -BeFalse
         $Global:Error | Should -BeNullOrEmpty
     }
 }