+++ CHANGELOG.md
@@ -1,2 +1,10 @@
 
 # Carbon.SmbShare Changelog
+
+## 1.0.0
+
+If migrating from Carbon:
+
+* Rename usages of `Install-CFileShare` to `Install-CSmbShare`.
+* Rename usages of `Test-CFileShare` to `Test-CSmbShare`.
+* Rename usages of `Uninstall-CFileShare` to `Uninstall-CSmbShare`.